+++ app/src/main/java/eu/faircode/email/AdapterMessage.java
@@ -4700,6 +4700,7 @@ public class AdapterMessage extends RecyclerView.Adapter<AdapterMessage.ViewHold
 
         private void onPickContact(String name, String email) {
             Intent pick = new Intent(Intent.ACTION_PICK, ContactsContract.Contacts.CONTENT_URI);
+            pick.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
             properties.setValue("name", name);
             properties.setValue("email", email);
             try {
@@ -5967,9 +5968,8 @@ public class AdapterMessage extends RecyclerView.Adapter<AdapterMessage.ViewHold
                         return null;
 
                     if (!TextUtils.isEmpty(message.inreplyto))
-                        for (String inreplyto : message.inreplyto.split(" "))
-                            for (EntityMessage m : db.message().getMessagesByMsgId(message.account, inreplyto))
-                                map.put(m.msgid, m);
+                        for (EntityMessage m : db.message().getMessagesByMsgId(message.account, message.inreplyto))
+                            map.put(m.msgid, m);
 
                     if (!TextUtils.isEmpty(message.references))
                         for (String ref : message.references.split(" "))
@@ -5994,8 +5994,7 @@ public class AdapterMessage extends RecyclerView.Adapter<AdapterMessage.ViewHold
                         start = ssb.length();
                         ssb.append("In-reply-to: ");
                         ssb.setSpan(new StyleSpan(Typeface.BOLD), start, ssb.length(), 0);
-                        for (String inreplyto : message.inreplyto.split(" "))
-                            ssb.append(inreplyto).append("\n");
+                        ssb.append(message.inreplyto).append("\n");
                     }
 
                     if (!TextUtils.isEmpty(message.references)) {