+++ app/src/main/java/eu/faircode/email/DaoMessage.java
@@ -647,6 +647,11 @@ public interface DaoMessage {
             " AND NOT ui_snoozed IS NULL")
     List<EntityMessage> getSnoozed(Long folder);
 
+    @Query("SELECT COUNT(*) FROM message" +
+            " WHERE NOT ui_snoozed IS NULL" +
+            " AND ui_snoozed <> " + Long.MAX_VALUE)
+    int getSnoozedCount();
+
     @Query("SELECT id AS _id, subject AS suggestion FROM message" +
             " WHERE (:account IS NULL OR message.account = :account)" +
             " AND (:folder IS NULL OR message.folder = :folder)" +