+++ app/src/main/java/eu/faircode/email/FragmentCompose.java
@@ -544,6 +544,7 @@ public class FragmentCompose extends FragmentBase {
 
                 // https://developer.android.com/guide/topics/providers/contacts-provider#Intents
                 Intent pick = new Intent(Intent.ACTION_PICK, ContactsContract.CommonDataKinds.Email.CONTENT_URI);
+                pick.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
                 startActivityForResult(Helper.getChooser(getContext(), pick), request);
             }
         };
@@ -6458,13 +6459,12 @@ public class FragmentCompose extends FragmentBase {
 
                         if (extras.getBoolean("archive")) {
                             EntityFolder archive = db.folder().getFolderByType(draft.account, EntityFolder.ARCHIVE);
-                            if (archive != null)
-                                for (String inreplyto : draft.inreplyto.split(" ")) {
-                                    List<EntityMessage> messages = db.message().getMessagesByMsgId(draft.account, inreplyto);
-                                    if (messages != null)
-                                        for (EntityMessage message : messages)
-                                            EntityOperation.queue(context, message, EntityOperation.MOVE, archive.id);
-                                }
+                            if (archive != null) {
+                                List<EntityMessage> messages = db.message().getMessagesByMsgId(draft.account, draft.inreplyto);
+                                if (messages != null)
+                                    for (EntityMessage message : messages)
+                                        EntityOperation.queue(context, message, EntityOperation.MOVE, archive.id);
+                            }
                         }
                     }
                 }
@@ -7952,15 +7952,18 @@ public class FragmentCompose extends FragmentBase {
                         return false;
                     }
 
-                    for (String inreplyto : draft.inreplyto.split(" ")) {
-                        List<EntityMessage> messages = db.message().getMessagesByMsgId(draft.account, inreplyto);
-                        for (EntityMessage message : messages) {
-                            EntityFolder folder = db.folder().getFolder(message.folder);
-                            if (folder == null)
-                                continue;
-                            if (EntityFolder.INBOX.equals(folder.type) || EntityFolder.USER.equals(folder.type))
-                                return true;
-                        }
+                    List<EntityMessage> messages = db.message().getMessagesByMsgId(draft.account, draft.inreplyto);
+                    if (messages == null || messages.size() == 0) {
+                        args.putString("reason", "In-reply-to gone");
+                        return false;
+                    }
+
+                    for (EntityMessage message : messages) {
+                        EntityFolder folder = db.folder().getFolder(message.folder);
+                        if (folder == null)
+                            continue;
+                        if (EntityFolder.INBOX.equals(folder.type) || EntityFolder.USER.equals(folder.type))
+                            return true;
                     }
 
                     args.putString("reason", "Not in inbox or unread");